Upstream version 1.1.0

# gpg: Unterschrift vom Di 10 Mär 2015 00:26:36 CET mittels RSA-Schlüssel ID 50605636
# gpg: Korrekte Unterschrift von "Sandro Knauß <bugs@sandroknauss.de>"

+++ tztable.h
@@ -420,5 +420,5 @@ static const char* olsonTimezones[] = {
     "Africa/Harare"
 };
 
-static const int numOlsonTimezones = sizeof olsonTimezones / sizeof *olsonTimezones;
+static const long unsigned int numOlsonTimezones = sizeof olsonTimezones / sizeof *olsonTimezones;
